Company 컴포넌트 변경, router 변경
