+++ src/pages/CompanyOld.module.css
@@ -1,0 +1,7 @@
+.ctt_con_p {
+  text-align: center;
+}
+
+.ctt_con_p_br {
+  clear: both;
+}
